Rename rgb_light to rgblight_user and fix all references to it
